requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.4.0
beautifulsoup4==4.11.1
better-profanity==0.7.0
bs4==0.0.1
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
cycler==0.11.0
decouple==0.0.7
discord==1.7.3
discord-components==2.1.2
discord-py-interactions==4.1.0
discord-py-slash-command==3.0.3
discord-ui==5.1.6
discord.py==1.7.3
ffmpeg-python==0.2.0
Flask==2.1.2
fonttools==4.32.0
future==0.18.2
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.2
MarkupSafe==2.1.1
matplotlib==3.5.1
multidict==6.0.2
nose==1.3.7
numpy==1.22.3
orjson==3.6.7
packaging==21.3
Pillow==9.0.1
pyparsing==3.0.7
python-dateutil==2.8.2
python-dotenv==0.20.0
PyYAML==6.0
Random-Word==1.0.7
requests==2.27.1
six==1.16.0
soupsieve==2.3.2.post1
typing_extensions==4.1.1
urllib3==1.26.9
Werkzeug==2.1.2
yarl==1.7.2
